upgrade wbtc-address-set to v3
