+++ .changeset/pink-waves-film.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/wbtc-address-set-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
