--- .changeset/six-pianos-grab.md
@@ -1,5 +1,0 @@
----
-'@chainlink/wbtc-address-set-test-adapter': major
----
-
-Initial version of EA
