+++ .pnp.cjs
@@ -780,10 +780,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
         "reference": "workspace:packages/sources/wbtc-address-set"\
       },\
       {\
-        "name": "@chainlink/wbtc-address-set-test-adapter",\
-        "reference": "workspace:packages/sources/wbtc-address-set-test"\
-      },\
-      {\
         "name": "@chainlink/wootrade-adapter",\
         "reference": "workspace:packages/sources/wootrade"\
       },\
@@ -1014,7 +1010,6 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
       ["@chainlink/vesper-adapter", ["workspace:packages/composites/vesper"]],\
       ["@chainlink/view-function-adapter", ["workspace:packages/sources/view-function"]],\
       ["@chainlink/wbtc-address-set-adapter", ["workspace:packages/sources/wbtc-address-set"]],\
-      ["@chainlink/wbtc-address-set-test-adapter", ["workspace:packages/sources/wbtc-address-set-test"]],\
       ["@chainlink/wootrade-adapter", ["workspace:packages/sources/wootrade"]],\
       ["@chainlink/wrapped-adapter", ["workspace:packages/sources/wrapped"]],\
       ["@chainlink/xbto-adapter", ["workspace:packages/sources/xbto"]],\
@@ -7191,6 +7186,25 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
             ["yeoman-generator", "npm:3.1.1"]\
           ],\
           "linkType": "HARD"\
+        }],\
+        ["npm:0.31.0", {\
+          "packageLocation": "./.yarn/cache/@chainlink-external-adapter-framework-npm-0.31.0-710681b9c8-9c8099e3b8.zip/node_modules/@chainlink/external-adapter-framework/",\
+          "packageDependencies": [\
+            ["@chainlink/external-adapter-framework", "npm:0.31.0"],\
+            ["ajv", "npm:8.11.0"],\
+            ["axios", "npm:1.3.4"],\
+            ["eventsource", "npm:2.0.2"],\
+            ["fastify", "npm:4.13.0"],\
+            ["ioredis", "npm:5.2.3"],\
+            ["mock-socket", "npm:9.1.5"],\
+            ["pino", "npm:8.6.1"],\
+            ["pino-pretty", "npm:9.1.0"],\
+            ["prom-client", "npm:13.2.0"],\
+            ["redlock", "npm:5.0.0-beta.2"],\
+            ["ws", "virtual:5aeb731ab2205b73ffebb216a4e187431be4ccd267893b7fe7611090ae6107a63d47ef052e0e827804831af35bc9bf8d2f1ce0d21f3432aef708b6225e3d245c#npm:8.9.0"],\
+            ["yeoman-generator", "npm:3.1.1"]\
+          ],\
+          "linkType": "HARD"\
         }]\
       ]],\
       ["@chainlink/external-adapters-js", [\
@@ -9140,26 +9154,7 @@ function $$SETUP_STATE(hydrateRuntimeState, basePath) {
           "packageLocation": "./packages/sources/wbtc-address-set/",\
           "packageDependencies": [\
             ["@chainlink/wbtc-address-set-adapter", "workspace:packages/sources/wbtc-address-set"],\
-            ["@chainlink/ea-bootstrap", "workspace:packages/core/bootstrap"],\
-            ["@chainlink/ea-test-helpers", "workspace:packages/core/test-helpers"],\
-            ["@chainlink/proof-of-reserves-adapter", "workspace:packages/composites/proof-of-reserves"],\
-            ["@types/jest", "npm:27.5.2"],\
-            ["@types/node", "npm:16.11.51"],\
-            ["@types/supertest", "npm:2.0.12"],\
-            ["nock", "npm:13.2.9"],\
-            ["supertest", "npm:6.2.4"],\
-            ["tslib", "npm:2.4.1"],\
-            ["typescript", "patch:typescript@npm%3A5.0.4#~builtin<compat/typescript>::version=5.0.4&hash=b5f058"]\
-          ],\
-          "linkType": "SOFT"\
-        }]\
-      ]],\
-      ["@chainlink/wbtc-address-set-test-adapter", [\
-        ["workspace:packages/sources/wbtc-address-set-test", {\
-          "packageLocation": "./packages/sources/wbtc-address-set-test/",\
-          "packageDependencies": [\
-            ["@chainlink/wbtc-address-set-test-adapter", "workspace:packages/sources/wbtc-address-set-test"],\
-            ["@chainlink/external-adapter-framework", "npm:0.30.1"],\
+            ["@chainlink/external-adapter-framework", "npm:0.31.0"],\
             ["@types/jest", "npm:27.5.2"],\
             ["@types/node", "npm:16.11.51"],\
             ["nock", "npm:13.2.9"],\
