--- packages/sources/wbtc-address-set-test/README.md
@@ -1,3 +1,0 @@
-# Chainlink External Adapter for wbtc-address-set-test
-
-This README will be generated automatically when code is merged to `main`. If you would like to generate a preview of the README, please run `yarn generate:readme wbtc-address-set-test`.
