--- packages/sources/wbtc-address-set-test/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@chainlink/wbtc-address-set-test-adapter",
-  "version": "0.0.0",
-  "description": "Chainlink wbtc-address-set-test adapter.",
-  "keywords": [
-    "Chainlink",
-    "LINK",
-    "blockchain",
-    "oracle",
-    "wbtc-address-set-test"
-  ],
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "repository": {
-    "url": "https://github.com/smartcontractkit/external-adapters-js",
-    "type": "git"
-  },
-  "license": "MIT",
-  "scripts": {
-    "clean": "rm -rf dist && rm -f tsconfig.tsbuildinfo",
-    "prepack": "yarn build",
-    "build": "tsc -b",
-    "server": "node -e 'require(\"./index.js\").server()'",
-    "server:dist": "node -e 'require(\"./dist/index.js\").server()'",
-    "start": "yarn server:dist"
-  },
-  "devDependencies": {
-    "@types/jest": "27.5.2",
-    "@types/node": "16.11.51",
-    "nock": "13.2.9",
-    "typescript": "5.0.4"
-  },
-  "dependencies": {
-    "@chainlink/external-adapter-framework": "0.30.1",
-    "tslib": "2.4.1"
-  }
-}
