--- packages/sources/wbtc-address-set-test/src/endpoint/index.ts
@@ -1,2 +1,0 @@
-export { endpoint as addresses } from './addresses'
-export { endpoint as members } from './members'
