--- packages/sources/wbtc-address-set-test/test-payload.json
@@ -1,3 +1,0 @@
-{
- "requests": [{}]
-}
