--- packages/sources/wbtc-address-set/schemas/env.json
@@ -1,21 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/wbtc-address-set-adapter.json",
-  "title": "Chainlink External Adapters for querying wBTC custodial address set",
-  "required": [],
-  "type": "object",
-  "properties": {
-    "MEMBERS_ENDPOINT": {
-      "type": "string",
-      "description": "wBTC endpoint of members (and their addresses). Required for members endpoint"
-    },
-    "ADDRESSES_ENDPOINT": {
-      "type": "string",
-      "description": "wBTC endpoint of addresses.  Required for addresses endpoint"
-    }
-  },
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
