+++ packages/sources/wbtc-address-set/src/config/index.ts
@@ -1,21 +1,12 @@
-import { Requester, util } from '@chainlink/ea-bootstrap'
-import { Config as BaseConfig } from '@chainlink/ea-bootstrap'
+import { AdapterConfig } from '@chainlink/external-adapter-framework/config'
 
-export const NAME = 'WBTC'
-
-export const DEFAULT_ENDPOINT = 'addresses'
-
-export const ENV_MEMBERS_ENDPOINT = 'MEMBERS_ENDPOINT'
-export const ENV_ADDRESSES_ENDPOINT = 'ADDRESSES_ENDPOINT'
-
-export type Config = BaseConfig & {
-  membersEndpoint?: string
-  addressesEndpoint?: string
-}
-
-export const makeConfig = (prefix = ''): Config => ({
-  ...Requester.getDefaultConfig(prefix),
-  defaultEndpoint: DEFAULT_ENDPOINT,
-  membersEndpoint: util.getEnv(ENV_MEMBERS_ENDPOINT, prefix),
-  addressesEndpoint: util.getEnv(ENV_ADDRESSES_ENDPOINT, prefix),
+export const config = new AdapterConfig({
+  MEMBERS_ENDPOINT: {
+    description: 'wBTC endpoint of members (and their addresses). Required for members endpoint',
+    type: 'string',
+  },
+  ADDRESSES_ENDPOINT: {
+    description: 'wBTC endpoint of addresses.  Required for addresses endpoint',
+    type: 'string',
+  },
 })
