+++ packages/sources/wbtc-address-set/src/endpoint/addresses.ts
@@ -1,49 +1,27 @@
-import { AdapterConfigError, Requester, Validator } from '@chainlink/ea-bootstrap'
-import { ExecuteWithConfig, InputParameters } from '@chainlink/ea-bootstrap'
-import { Config } from '../config'
-import { PorInputAddress } from '@chainlink/proof-of-reserves-adapter/src/utils/PorInputAddress'
+import { config } from '../config'
+import { httpTransport } from '../transport/addresses'
+import { EmptyInputParameters } from '@chainlink/external-adapter-framework/validation/input-params'
+import { AdapterInputError } from '@chainlink/external-adapter-framework/validation/error'
+import {
+  PoRAddressEndpoint,
+  PoRAddressResponse,
+} from '@chainlink/external-adapter-framework/adapter/por'
 
-export const supportedEndpoints = ['addresses']
-
-type APIMembersResponse = {
-  result: Address[]
-  count: number
+export type BaseEndpointTypes = {
+  Parameters: EmptyInputParameters
+  Response: PoRAddressResponse
+  Settings: typeof config.settings
 }
 
-type Address = {
-  id: string
-  address: string
-  balance?: number
-  type: AddressType
-  verified: boolean
-}
-
-type AddressType = 'custodial' | 'merchant' | 'deposit'
-export type TInputParameters = Record<string, never>
-export const inputParameters: InputParameters<TInputParameters> = {}
-
-export const execute: ExecuteWithConfig<Config> = async (request, _, config) => {
-  const validator = new Validator(request, inputParameters)
-
-  Requester.logConfig(config)
-
-  const jobRunID = validator.validated.id
-
-  if (!config.addressesEndpoint) {
-    throw new AdapterConfigError({
-      jobRunID,
-      statusCode: 500,
-      message: 'The address list endpoint has not been configured for this adapter',
-    })
-  }
-
-  const options = { ...config.api, baseURL: config.addressesEndpoint }
-  const response = await Requester.request<APIMembersResponse>(options)
-
-  const result = response.data.result
-    .filter((a) => a.type == 'custodial' && a.balance)
-    .map<PorInputAddress>((a) => ({ ...a, coin: 'btc', chainId: 'mainnet', network: 'bitcoin' }))
-
-  const output = { ...response, data: { ...response.data, result } }
-  return Requester.success(jobRunID, output, config.verbose)
-}
+export const endpoint = new PoRAddressEndpoint({
+  name: 'addresses',
+  transport: httpTransport,
+  customInputValidation: (_, adapterSettings): AdapterInputError | undefined => {
+    if (!adapterSettings.ADDRESSES_ENDPOINT) {
+      throw new AdapterInputError({
+        message: `ADDRESSES_ENDPOINT env var is required for 'addresses' endpoint`,
+      })
+    }
+    return
+  },
+})
