+++ packages/sources/wbtc-address-set/src/endpoint/index.ts
@@ -1,7 +1,2 @@
-import type { TInputParameters as AddressesInputParameters } from './addresses'
-import type { TInputParameters as MembersInputParameters } from './members'
-
-export type TInputParameters = AddressesInputParameters | MembersInputParameters
-
-export * as addresses from './addresses'
-export * as members from './members'
+export { endpoint as addresses } from './addresses'
+export { endpoint as members } from './members'
