+++ packages/sources/wbtc-address-set/src/index.ts
@@ -1,10 +1,13 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { endpointSelector, makeExecute } from './adapter'
-import * as endpoints from './endpoint'
-import { makeConfig, NAME } from './config'
-import * as types from './endpoint'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { PoRAdapter } from '@chainlink/external-adapter-framework/adapter/por'
+import { config } from './config'
+import { addresses, members } from './endpoint'
 
-const adapterContext = { name: NAME }
+export const adapter = new PoRAdapter({
+  defaultEndpoint: addresses.name,
+  name: 'WBTC-ADDRESS-SET-TEST',
+  config,
+  endpoints: [addresses, members],
+})
 
-const { server } = expose(adapterContext, makeExecute(), undefined, endpointSelector)
-export { NAME, makeExecute, makeConfig, server, types, endpoints }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
