+++ packages/sources/wbtc-address-set/test/integration/adapter.test.ts
@@ -1,65 +1,51 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import * as process from 'process'
-import { server as startServer } from '../../src'
+import {
+  TestAdapter,
+  setEnvVariables,
+} from '@chainlink/external-adapter-framework/util/testing-utils'
+import * as nock from 'nock'
 import { mockAddressesResponseSuccess, mockMembersResponseSuccess } from './fixtures'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
-import { SuperTest, Test } from 'supertest'
 
 describe('execute', () => {
-  const id = '1'
-  const context: SuiteContext = {
-    req: null,
-    server: startServer,
-  }
+  let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
 
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-    MEMBERS_ENDPOINT: process.env.MEMBERS_ENDPOINT || 'http://localhost:8081',
-    ADDRESSES_ENDPOINT: process.env.ADDRESSES_ENDPOINT || 'http://localhost:8082',
-  }
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    process.env.ADDRESSES_ENDPOINT = process.env.ADDRESSES_ENDPOINT ?? 'http://localhost:8082'
+    process.env.MEMBERS_ENDPOINT = process.env.MEMBERS_ENDPOINT ?? 'http://localhost:8081'
+    const mockDate = new Date('2001-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
 
-  setupExternalAdapterTest(envVariables, context)
-
-  describe('members api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
-        endpoint: 'members',
-      },
-    }
-
-    it('should return success', async () => {
-      mockMembersResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+    const adapter = (await import('./../../src')).adapter
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
     })
   })
 
-  describe('addresses api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {},
-    }
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
+    spy.mockRestore()
+  })
 
+  describe('addresses endpoint', () => {
     it('should return success', async () => {
       mockAddressesResponseSuccess()
+      const response = await testAdapter.request({})
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
+    })
+  })
 
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+  describe('members endpoint', () => {
+    it('should return success', async () => {
+      mockMembersResponseSuccess()
+      const response = await testAdapter.request({ endpoint: 'members' })
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 })
