+++ packages/tsconfig.test.json
@@ -564,9 +564,6 @@
       "path": "./sources/wbtc-address-set/tsconfig.test.json"
     },
     {
-      "path": "./sources/wbtc-address-set-test/tsconfig.test.json"
-    },
-    {
       "path": "./sources/wootrade/tsconfig.test.json"
     },
     {
